Thedev

+++ app/Http/Controllers/API/CategoryController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers\API;
 
-use App\Http\Controllers\Controller;
+use App\Models\Role;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -26,6 +28,13 @@ class CategoryController extends Controller
     public function store(Request $request)
     {
         try {
+            $authenticatedUser = Auth::user();
+
+            // Check if the user has the required role
+            if (!$authenticatedUser->roleUser || $authenticatedUser->roleUser->role_id !== Role::ROLE_ADMIN) {
+                throw new \Exception('You are not authorized to create a category.');
+            }
+
             $validatedData = $request->validate([
                 'name' => 'required',
                 'slug' => 'required|unique:categories',
@@ -51,6 +60,13 @@ class CategoryController extends Controller
     public function update(Request $request, Category $category)
     {
         try {
+            $authenticatedUser = Auth::user();
+
+            // Check if the user has the required role
+            if (!$authenticatedUser->roleUser || $authenticatedUser->roleUser->role_id !== Role::ROLE_ADMIN) {
+                throw new \Exception('You are not authorized to update this category.');
+            }
+
             $validatedData = $request->validate([
                 'name' => 'required',
                 'slug' => 'required|unique:categories,slug,' . $category->id,
@@ -62,10 +78,17 @@ class CategoryController extends Controller
             return response()->json(['success' => false, 'message' => $e->getMessage()]);
         }
     }
-
+    
     public function destroy(Category $category)
     {
         try {
+            $authenticatedUser = Auth::user();
+
+            // Check if the user has the required role
+            if (!$authenticatedUser->roleUser || $authenticatedUser->roleUser->role_id !== Role::ROLE_ADMIN) {
+                throw new \Exception('You are not authorized to delete this category.');
+            }
+
             $category->delete();
             return response()->json(['success' => true], 204);
         } catch (\Exception $e) {
